// expo/types/game.ts
export interface Race {
  id: string;
  name: string;
  description: string;
  statBonuses: {
    [key: string]: number;
  };
  abilities: string[];
  lore: string;
}
// ...
export interface CustomRace {
  name: string;
  description: string;
  statBonuses: {
    [stat: string]: number;
  };
  abilities: string[];
  lore: string;
}